web-230-MapDraweSearchRefactor: new drawer created and setup on map page above bottom drawer
