+++ compnents/FABMenu/siteSearchButton.tsx
@@ -46,7 +46,7 @@ export default function SiteSearchButton() {
 
 const styles = StyleSheet.create({
   container: {
-    // backgroundColor: "white",
+    // backgroundColor: "green",
     opacity: 1,
     alignItems: "center",
     justifyContent: "center",
